requirements.txt
requests>=2.31.0
twilio>=8.11.0
python-dotenv>=1.0.0
pytz>=2023.3
tenacity>=8.2.0
openai>=1.0.0